packages/sdk/output/typescript/src/services/organisations/Organisations.ts
import BaseService from '../../BaseService';

import { CreateOrganizationDto } from './models/CreateOrganizationDto';

export class OrganisationsService extends BaseService {
  async organisationsControllerGetOragnisations(): Promise<any> {
    const urlEndpoint = '/organisations';
    const finalUrl = encodeURI(`${this.baseUrl + urlEndpoint}`);
    const response: any = await this.httpClient.get(
      finalUrl,
      {},
      {
        ...this.getAuthorizationHeader(),
      },
      true,
    );
    const responseModel = response.data;
    return responseModel;
  }

  async organisationsControllerCreateOrg(input: CreateOrganizationDto): Promise<any> {
    const headers: { [key: string]: string } = { 'Content-type': 'application/json' };
    const urlEndpoint = '/organisations/create';
    const finalUrl = encodeURI(`${this.baseUrl + urlEndpoint}`);
    const response: any = await this.httpClient.post(
      finalUrl,
      input,
      {
        ...headers,
        ...this.getAuthorizationHeader(),
      },
      true,
    );
    const responseModel = response.data;
    return responseModel;
  }
}
